third_party/WebKit/LayoutTests/fast/forms/autofilled-expected.txt

This tests that foreground and background colors properly change for autofilled inputs or select options. It can only be run using the test harness.
           
PASS successfullyParsed is true

TEST COMPLETE
PASS foregroundOf(textarea1) is originalForeground
PASS backgroundOf(textarea1) is originalBackground
PASS search.value is "Search value"
PASS foregroundOf(field) is not originalForeground
PASS backgroundOf(field) is not originalBackground
PASS foregroundOf(search) is not originalForeground
PASS backgroundOf(search) is not originalBackground
PASS foregroundOf(textarea1) is not originalForeground
PASS backgroundOf(textarea1) is not originalBackground
PASS foregroundOf(textarea2) is not originalForeground
PASS backgroundOf(textarea2) is not originalBackground
PASS foregroundOf(select1) is not originalForeground
PASS backgroundOf(select1) is not originalBackground
PASS foregroundOf(select2) is not originalForeground
PASS backgroundOf(select2) is not originalBackground
PASS foregroundOf(select3) is not originalForeground
PASS backgroundOf(select3) is not originalBackground
PASS select2.options.length is 3
PASS select2.options.length is 2
PASS foregroundOf(select2) is autofilledSelectForeground
PASS backgroundOf(select2) is autofilledSelectBackground
PASS search.value is ""
PASS foregroundOf(field) is originalForeground
PASS backgroundOf(field) is originalBackground
PASS foregroundOf(search) is originalForeground
PASS backgroundOf(search) is originalBackground
PASS foregroundOf(textarea1) is originalForeground
PASS backgroundOf(textarea1) is originalBackground
PASS foregroundOf(textarea2) is originalForeground
PASS backgroundOf(textarea2) is originalBackground
PASS foregroundOf(select1) is originalForeground
PASS backgroundOf(select1) is originalBackground
PASS foregroundOf(select2) is originalForeground
PASS backgroundOf(select2) is originalBackground
PASS foregroundOf(select3) is originalForeground
PASS backgroundOf(select3) is originalBackground
